refactor(Jumpseller::Spider): add parent class
